Move text reading functions into basic.py
